refactor(map): optimize component, maximize display, animation
- Expand map to full width when toggled
- Add animation
- Externalize toggle state (separate button)
